feat: grouped settings
