+++ src/settings/settings.base.ts
@@ -3,11 +3,16 @@ import { NutstoreSettingTab } from '.'
 import NutstorePlugin from '..'
 
 export default abstract class BaseSettings {
+	/** Dynamic because translated labels may change while the settings pane is open. */
+	abstract readonly name: () => string
+	readonly searchable: boolean = true
+	abstract readonly showGroupHeading: boolean
+
 	constructor(
 		protected app: App,
 		protected plugin: NutstorePlugin,
 		protected settings: NutstoreSettingTab,
-		protected containerEl: HTMLElement,
+		public readonly containerEl: HTMLElement,
 	) {}
 
 	/** Public control labels only; indexing must not render controls or read user values. */
